refactor(modules): rewrite from CJS to ESM
